requirements.txt
cython
numpy
albumentations
imagecorruptions
mmcv
matplotlib
terminaltables
six
pycocotools
pillow==6.2.2
shapely
ensemble-boxes
cairosvg
svgpathtools
svgutils
shapely